Moving 'check for updates' to general settings
