+++ src/client/settings/Settings.tsx
@@ -1,7 +1,6 @@
 import React, {ReactElement} from 'react'
 import {GeneralSettings} from './GeneralSettings'
 import {DisplaySettings} from './DisplaySettings'
-import {MiscellaneousSettings} from './MiscellaneousSettings'
 import {Title} from '../common/Title'
 import {NotificationSettings} from './NotificationSettings'
 
@@ -12,7 +11,6 @@ export function Settings(): ReactElement {
       <GeneralSettings/>
       <DisplaySettings/>
       <NotificationSettings/>
-      <MiscellaneousSettings/>
     </>
   )
 }
